Split command line code out from API.
